+++ .idea/dictionaries/project.xml
@@ -2,6 +2,7 @@
   <dictionary name="project">
     <words>
       <w>desplazamiento</w>
+      <w>direccion</w>
       <w>posicion</w>
     </words>
   </dictionary>
