+++ src/main/java/org/modelo/ArmaUnDisparo.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 public class ArmaUnDisparo {
     private Disparo disparoEnCurso;
