+++ src/main/java/org/modelo/BloqueBosque.java
@@ -1,13 +1,13 @@
-package org.example;
+package org.modelo;
 
-public class BloqueAgua extends Bloque{
-    public BloqueAgua(Vector2D posicion) {
+public class BloqueBosque extends Bloque {
+    public BloqueBosque(Vector2D posicion) {
         super(posicion.obtenerCoordenadaX(), posicion.obtenerCoordenadaY());
     }
 
     @Override
     public boolean bloqueaPasoTanque() {
-        return true;
+        return false;
     }
 
     @Override
@@ -24,4 +24,9 @@ public class BloqueAgua extends Bloque{
     public void recibeimpacto() {
         // No hace nada
     }
+
+    public boolean ocultaTanque() {
+        return true; // la vista puede usar esto para dibujar tanques ocultos
+    }
+
 }
