+++ src/main/java/org/modelo/Direccion.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 public enum Direccion {
     ARRIBA,
