+++ src/main/java/org/modelo/Disparo.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 public class Disparo {
     private Vector2D posicion;
