+++ src/main/java/org/modelo/Enemigo.java
@@ -1,5 +1,5 @@
 
-package org.example;
+package org.modelo;
 
 public class Enemigo {
     private Tanque tanque;
