+++ src/main/java/org/modelo/EstadoNivel.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 public enum EstadoNivel {
     EN_CURSO, VICTORIA, DERROTA;
