+++ src/main/java/org/modelo/Main.java
@@ -1,6 +1,4 @@
-package org.example;
-
-import java.io.InputStream;
+package org.modelo;
 
 public class Main {
     public static void main(String[] args) {
