+++ src/main/java/org/modelo/Nivel.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 import java.util.ArrayList;
 import java.util.List;
