+++ src/main/java/org/modelo/Vector2D.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 public class Vector2D {
