+++ src/test/java/org/modelo/JugadorTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
