+++ src/test/java/org/modelo/NivelTest.java
@@ -1,8 +1,5 @@
-package org.example;
+package org.modelo;
 
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 /*
 class NivelTest {
