+++ src/test/java/org/modelo/PosicionTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 import org.junit.jupiter.api.Test;
