+++ src/test/java/org/modelo/Vector2DTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.modelo;
 
 import org.junit.jupiter.api.Test;
